Make error message available
Instead of forcing us to show same error message. It would be better to allow the server to provide that message.
I saw your example in `nodejs` version and there, your provide custom error message for each error, but in `client` side, you did not make use of that errors.